+++ src/parse.rs
@@ -11,21 +11,20 @@ pub fn languages(input: &str) -> Vec<Language> {
 
             let name = lang.get("name").unwrap().as_str().unwrap().to_string();
 
-            let file_types = lang
-                .get("file-types")
-                .unwrap()
-                .as_array()
-                .unwrap()
-                .iter()
-                .map(|file_type| {
-                    if let Some(file_type) = file_type.as_str() {
-                        file_type.to_string()
-                    } else {
-                        let file_type = file_type.as_table().unwrap().get("glob").unwrap();
-                        file_type.as_str().unwrap().to_string()
-                    }
-                })
-                .collect();
+            let file_types = lang.get("file-types").map(|ft| {
+                ft.as_array()
+                    .unwrap()
+                    .iter()
+                    .map(|file_type| {
+                        if let Some(file_type) = file_type.as_str() {
+                            file_type.to_string()
+                        } else {
+                            let file_type = file_type.as_table().unwrap().get("glob").unwrap();
+                            file_type.as_str().unwrap().to_string()
+                        }
+                    })
+                    .collect()
+            });
 
             let (size, style) = if let Some(indent) = lang.get("indent") {
                 let size = indent
@@ -50,13 +49,12 @@ pub fn languages(input: &str) -> Vec<Language> {
 
             Language {
                 name,
-                file_types,
                 cfg: LangCfg {
                     size,
                     style,
                     tab_width: None,
                     max_line_length: None,
-                    file_types: vec![],
+                    file_types,
                 },
             }
         })